+++ docs/versionhistory.rst
@@ -13,7 +13,6 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
   * ``current_time()``
   * ``current_effective_deadline()``
   * ``MemoryObjectReceiveStream.receive_nowait()``
-  * ``MemoryObjectSendStream.send_nowait()``
 
   The following functions and methods are no longer asynchronous but can still be awaited on (doing
   so will emit a deprecation warning):
@@ -25,6 +24,7 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
   * ``CapacityLimiter.acquire_on_behalf_of_nowait()``
   * ``Condition.release()``
   * ``Event.set()``
+  * ``MemoryObjectSendStream.send_nowait()``
   * ``Lock.release()``
   * ``Semaphore.release()``
   * ``TaskGroup.spawn()``